adds shift parameters, expose templates, test tuning
