+++ clients/gtest/blas1_gtest.yaml
@@ -82,18 +82,39 @@ Tests:
 
 - name: blas1_asum_batched
   category: quick
-  N: [ -1, 0, 5, 10, 1025, 7111, 10000, 33792 ]
+  N: [ -1, 0, 1025 ]
+  batch_count: [-1, 0, 1]
+  function: 
+    - asum_batched: *single_double_precisions_complex_real
+- name: blas1_asum_batched
+  category: pre_checkin
+  N: [ 1025, 33792 ]
   incx: [-1, 1, 2]
-  batch_count: [-1, 1, 2, 10, 1025]
+  batch_count: [ 2, 33 ]
   function: 
     - asum_batched: *single_double_precisions_complex_real
 
 - name: blas1_asum_strided_batched
   category: quick
-  N: [ -1, 0, 5, 10, 1025, 7111, 10000, 33792 ]
+  N: [ -1, 0, 1025 ]
+  stride_x: [-1, 68000]
+  batch_count: [-1, 0, 1]
+  function: 
+    - asum_strided_batched: *single_double_precisions_complex_real
+- name: blas1_asum_strided_batched
+  category: pre_checkin
+  N: [ 1025, 33792 ]
   incx: [-1, 1, 2]
   stride_x: [-1, 68000]
-  batch_count: [-1, 1, 2, 10, 1025]
+  batch_count: [ 2, 33 ]
+  function: 
+    - asum_strided_batched: *single_double_precisions_complex_real
+- name: blas1_asum_strided_batched
+  category: nightly
+  N: [ 8000001 ]
+  incx: [-1, 1, 2]
+  stride_x: [-1, 16000002]
+  batch_count: [ 2, 1025 ]
   function: 
     - asum_strided_batched: *single_double_precisions_complex_real
